fix(encoder): handle non-exhaustive ColorRange and fix a second den/num bug
Two more real compile errors from the same first-ever CI compile:

- ColorRange is #[non_exhaustive] (declared in mediaway-common) so a
  same-purpose match in mediaway-encoder needs a wildcard arm even though
  both real variants are covered; map an unknown future variant to
  EncodeError::Unsupported rather than a panic.
- The Android AMediaCodec encoder's frame_rate_hint had the identical
  u32/u64 Rational division bug already fixed in the Apple encoder's copy
  of this helper — same fix (widen den to u64 before dividing by num).

+++ crates/mediaway-encoder/src/android/amediacodec/video.rs
@@ -265,7 +265,7 @@ fn frame_rate_hint(time_base: mediaway_common::Rational) -> i32 {
     if time_base.num == 0 {
         return 30;
     }
-    i32::try_from(time_base.den / time_base.num)
+    i32::try_from(u64::from(time_base.den) / time_base.num)
         .unwrap_or(30)
         .max(1)
 }
